JS-201: Replace the store
