+++ src/index.js
@@ -1,10 +1,11 @@
 import './main.css';
 
-import store from './store';
+import todoStore from './todoStore';
 import render from './render';
 import './registerEventListeners';
 
+
 const rootElement = document.getElementById('demo');
 
-store.subscribe(state => render(rootElement, state));
-render(rootElement, store.getState());
+todoStore.subscribe(() => render(rootElement, todoStore.getState()))
+render(rootElement, todoStore.getState());
