--- src/store/index.js
@@ -1,4 +1,0 @@
-import createStore from './createStore';
-import todoReducer, { initialState as todoInitialState } from './todos';
-
-export default createStore(todoReducer, todoInitialState);
