--- src/store/todoActions.js
@@ -1,9 +1,0 @@
-export const toggle = id => ({
-  type: 'TOGGLE',
-  id
-});
-
-export const add = title => ({
-  type: 'ADD',
-  title
-});
